+++ .gitignore
@@ -31,3 +31,6 @@ packages/create-atomic/template/package.json.hbs
 
 # Prevent secret leakage.
 .npmrc
+
+# Commit message file used for release
+.git-message
