+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "cli-tools",
   "private": true,
-  "version": "1.35.12",
+  "version": "1.0.0-0",
   "author": "Coveo",
   "bin": {
     "coveo": "./bin/run"
@@ -84,11 +84,10 @@
     "commit-msg": "node ./hooks/commit-msg.js",
     "pre-commit": "lint-staged",
     "postinstall": "patch-package && nx run-many --target=postinstall --all",
-    "release": "npm run nx:graph && npm run release:phase1 && npm run release:phase2 && npm run release:phase3",
+    "release": "npm run nx:graph && npm run release:phase1 && npm run release:phase2",
     "nx:graph": "nx graph --file=topology.json",
-    "release:phase1": "node --experimental-specifier-resolution=node ./scripts/releaseV2/phase1-get-new-version.mjs",
-    "release:phase2": "nx run-many --target=release:phase2 --all --parallel=false",
-    "release:phase3": "node --experimental-specifier-resolution=node ./scripts/releaseV2/phase3-git-commit-tag-push.mjs"
+    "release:phase1": "nx run-many --target=release:phase1 --all --parallel=false",
+    "release:phase2": "node --experimental-specifier-resolution=node ./scripts/releaseV2/phase2-git-commit-tag-push.mjs"
   },
   "workspaces": [
     "packages/*",
