+++ packages/atomic/template/package.json
@@ -4,7 +4,7 @@
   "version": "1.35.12",
   "private": true,
   "scripts": {
-    "release:phase2": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase2-bump-all-packages.mjs",
+    "release:phase1": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase1-bump-all-packages.mjs",
     "lint": "prettier --check . && eslint ."
   },
   "dependencies": {
