+++ packages/cli/commons/package.json
@@ -36,7 +36,7 @@
     "test": "jest",
     "build": "tsc -p tsconfig.build.json",
     "prepublishOnly": "npm run build",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase2-bump-all-packages.mjs"
+    "release:phase1": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase1-bump-all-packages.mjs"
   },
   "publishConfig": {
     "access": "public"
