+++ packages/cli/core/package.json
@@ -11,13 +11,13 @@
     "@amplitude/node": "^1.9.0",
     "@coveo/cli-commons": "1.35.12",
     "@coveo/cli-plugin-source": "1.35.12",
-    "@coveo/push-api-client": "^2.4.4",
+    "@coveo/push-api-client": "2.5.8",
     "@coveord/platform-client": "34.13.0",
     "@oclif/core": "1.16.3",
-    "@oclif/plugin-help": "^5",
-    "@oclif/plugin-plugins": "^2.1.0",
-    "@oclif/plugin-update": "^3.0.0",
-    "@oclif/plugin-version": "^1.0.4",
+    "@oclif/plugin-help": "5.1.12",
+    "@oclif/plugin-plugins": "2.1.0",
+    "@oclif/plugin-update": "3.0.0",
+    "@oclif/plugin-version": "1.1.2",
     "abortcontroller-polyfill": "^1.7.1",
     "archiver": "^5.3.0",
     "async-retry": "^1.3.1",
@@ -31,13 +31,13 @@
     "is-unicode-supported": "^1.3.0",
     "isomorphic-fetch": "^3.0.0",
     "json2csv": "^5.0.6",
-    "jsonschema": "^1.4.0",
+    "jsonschema": "1.4.1",
     "node-machine-id": "^1.1.12",
     "open": "^8.4.0",
     "semver": "^7.3.4",
     "tmp": "^0.2.1",
-    "ts-dedent": "^2.1.1",
-    "tslib": "^2"
+    "ts-dedent": "2.2.0",
+    "tslib": "2.4.0"
   },
   "devDependencies": {
     "@amplitude/types": "1.10.2",
@@ -70,7 +70,7 @@
     "strip-ansi": "6.0.1",
     "ts-jest": "27.1.5",
     "ts-node": "10.9.1",
-    "tsconfig-paths": "3.14.1",
+    "tsconfig-paths": "4.1.0",
     "typescript": "4.8.3"
   },
   "peerDependencies": {
@@ -175,7 +175,7 @@
     "build": "rimraf lib && tsc -p tsconfig.build.json",
     "test": "jest --colors",
     "lint": "prettier --check . && eslint .",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase2-bump-all-packages.mjs",
+    "release:phase1": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase1-bump-all-packages.mjs",
     "postpack": "rimraf oclif.manifest.json",
     "prepublishOnly": "rimraf lib && npm run build && oclif manifest && oclif readme"
   },
