+++ packages/cli/source/package.json
@@ -36,7 +36,7 @@
     "strip-ansi": "6.0.1",
     "ts-jest": "27.1.5",
     "ts-node": "10.9.1",
-    "tsconfig-paths": "3.14.1",
+    "tsconfig-paths": "4.1.0",
     "tslib": "2.4.0",
     "typescript": "4.8.3"
   },
@@ -67,7 +67,7 @@
     "build": "tsc -p tsconfig.build.json",
     "test": "jest --colors",
     "lint": "prettier --config .prettierrc.js --check . && eslint .",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase2-bump-all-packages.mjs",
+    "release:phase1": "node --experimental-specifier-resolution=node ../../../scripts/releaseV2/phase1-bump-all-packages.mjs",
     "postpack": "rimraf oclif.manifest.json",
     "prepublishOnly": "rimraf lib && npm run build && oclif manifest"
   },
