+++ packages/cra-template/package.json
@@ -3,7 +3,7 @@
   "version": "1.35.12",
   "scripts": {
     "lint": "prettier --check . && eslint .",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase2-bump-all-packages.mjs"
+    "release:phase1": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase1-bump-all-packages.mjs"
   },
   "keywords": [
     "react",
