+++ packages/search-token-lambda/package.json
@@ -12,7 +12,7 @@
     "start": "netlify functions:serve",
     "test": "jest",
     "test:watch": "jest --watch",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase2-bump-all-packages.mjs"
+    "release:phase1": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase1-bump-all-packages.mjs"
   },
   "dependencies": {
     "@coveord/platform-client": "34.13.0",
