+++ packages/search-token-server/package.json
@@ -25,7 +25,7 @@
   "scripts": {
     "start": "ts-node server.ts",
     "test": "jest",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase2-bump-all-packages.mjs"
+    "release:phase1": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase1-bump-all-packages.mjs"
   },
   "devDependencies": {
     "@babel/core": "7.19.1",
