+++ packages/vue-cli-plugin-typescript/package.json
@@ -10,7 +10,7 @@
   "license": "Apache-2.0",
   "scripts": {
     "lint": "prettier --check . && eslint .",
-    "release:phase2": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase2-bump-all-packages.mjs"
+    "release:phase1": "node --experimental-specifier-resolution=node ../../scripts/releaseV2/phase1-bump-all-packages.mjs"
   },
   "devDependencies": {
     "@coveo/headless": "1.103.5",
